refactored newsletters
