+++ apps/devportal/src/lib/newsletter.ts
@@ -1,2 +1,40 @@
-export const getNewsletterTitle = (dateAsYearMonth: string, title?: string) =>
-  title ? `${title} - ${dateAsYearMonth}` : dateAsYearMonth;
+import { NEWSLETTER_DATA_DIRECTORY } from '@lib/staticPaths';
+import fs from 'fs';
+import path from 'path';
+import { translateDateAsYearMonth } from 'ui/lib/utils/dateUtil';
+
+const MAX_RESULTS = 12;
+
+export const getNewsletterTitle = (dateAsYearMonth: string, title?: string) => (title ? `${title} - ${dateAsYearMonth}` : dateAsYearMonth);
+
+export const getFirstXNewsletters = () => {
+  const years = fs.readdirSync(NEWSLETTER_DATA_DIRECTORY);
+
+  const newsletters = [];
+
+  years.sort().reverse();
+
+  // Using for loops to shortcut early
+  for (let i = 0; i < years.length; i++) {
+    const year = years[i];
+    const yearPath = path.resolve(NEWSLETTER_DATA_DIRECTORY, `${year}`);
+    const months = fs.readdirSync(yearPath).sort().reverse();
+    for (let j = 0; j < months.length; j++) {
+      const month = months[j];
+      const { title, description } = JSON.parse(fs.readFileSync(path.resolve(yearPath, `${month}`), { encoding: 'utf-8' }));
+
+      const monthWithoutFile = month.substring(0, 2);
+      newsletters.push({
+        title: getNewsletterTitle(translateDateAsYearMonth(`${year}-${monthWithoutFile}-03`), title),
+        description,
+        href: `newsletter/${year}/${monthWithoutFile}`,
+      });
+
+      if (newsletters.length === MAX_RESULTS) {
+        return newsletters;
+      }
+    }
+  }
+
+  return newsletters;
+};
